+++ tum/computational-fluid-dynamics/lbm-gpu/src/boundary.h
@@ -2,6 +2,6 @@
 #define _BOUNDARY_H_
 
 /** handles the boundaries in our simulation setup */
-void TreatBoundary(double *collide_field, int* flag_field, const double * const wall_velocity,
+void TreatBoundary(float *collide_field, int* flag_field, const float * const wall_velocity,
         int xlength);
 #endif
