+++ tum/computational-fluid-dynamics/lbm-gpu/src/collision.h
@@ -4,6 +4,6 @@
 /** carries out the whole local collision process. Computes density and velocity and
  *  equilibrium distributions. Carries out BGK update.
  */
-void DoCollision(double *collide_field, int *flag_field, double tau, int xlength);
+void DoCollision(float *collide_field, int *flag_field, float tau, int xlength);
 
 #endif
