+++ tum/computational-fluid-dynamics/lbm-gpu/src/collision_gpu.cu
@@ -18,7 +18,7 @@ inline void cudaAssert(cudaError_t code, char *file, int line, bool abort=true){
 }
 
 
-__constant__ double tau_d;
+__constant__ float tau_d;
 __constant__ int xlength_d;
 //TODO:move lattice constants to gpu constant memory
 __device__ static const int LATTICE_VELOCITIES_D[19][3] = {
@@ -26,7 +26,7 @@ __device__ static const int LATTICE_VELOCITIES_D[19][3] = {
     {-1,0,0}, {0,0,0},  {1,0,0}, {-1,1,0},{0,1,0}, {1,1,0},  {0,-1,1},{-1,0,1},
     {0,0,1},  {1,0,1},  {0,1,1}
 };
-__device__ static const double LATTICE_WEIGHTS_D[19] = {
+__device__ static const float LATTICE_WEIGHTS_D[19] = {
     1.0/36.0, 1.0/36.0, 2.0/36.0, 1.0/36.0, 1.0/36.0, 1.0/36.0, 2.0/36.0, 1.0/36.0,
     2.0/36.0, 12.0/36.0,2.0/36.0, 1.0/36.0, 2.0/36.0, 1.0/36.0, 1.0/36.0, 1.0/36.0,
     2.0/36.0, 1.0/36.0, 1.0/36.0
@@ -38,7 +38,7 @@ __device__ static const double LATTICE_WEIGHTS_D[19] = {
  * currentCell thus denotes the address of the first particle distribution function of the
  * respective cell. The result is stored in density.
  */
-__device__ void ComputeDensity(double *current_cell, double *density){
+__device__ void ComputeDensity(float *current_cell, float *density){
     int i; *density=0;
     for(i=0;i<Q_LBM;i++)
         *density+=current_cell[i];
@@ -52,7 +52,7 @@ __device__ void ComputeDensity(double *current_cell, double *density){
 /**
  * Computes the velocity within currentCell and stores the result in velocity
  */
-__device__ void ComputeVelocity(double *current_cell, double *density, double *velocity){
+__device__ void ComputeVelocity(float *current_cell, float *density, float *velocity){
     int i;
     velocity[0]=0;
     velocity[1]=0;
@@ -74,9 +74,9 @@ __device__ void ComputeVelocity(double *current_cell, double *density, double *v
  * Computes the equilibrium distributions for all particle distribution functions of one
  * cell from density and velocity and stores the results in feq.
  */
-__device__ void ComputeFeq(double *density, double *velocity, double *feq){
+__device__ void ComputeFeq(float *density, float *velocity, float *feq){
     int i;
-    double s1, s2, s3;
+    float s1, s2, s3;
     for(i=0;i<Q_LBM;i++){
         s1 = LATTICE_VELOCITIES_D[i][0]*velocity[0]+LATTICE_VELOCITIES_D[i][1]*velocity[1]+
         		LATTICE_VELOCITIES_D[i][2]*velocity[2];
@@ -96,7 +96,7 @@ __device__ void ComputeFeq(double *density, double *velocity, double *feq){
  * Computes the post-collision distribution functions according to the BGK update rule and
  * stores the results again at the same position.
  */
-__device__ void ComputePostCollisionDistributions(double *current_cell, double *feq){
+__device__ void ComputePostCollisionDistributions(float *current_cell, float *feq){
     int i;
     for(i=0;i<Q_LBM;i++){
         current_cell[i]=current_cell[i]-(current_cell[i]-feq[i])/tau_d;
@@ -110,9 +110,9 @@ __device__ void ComputePostCollisionDistributions(double *current_cell, double *
 /**
  * Performs the actual collision computation
  */
-__global__ void DoColision(double *collide_field_d){
+__global__ void DoColision(float *collide_field_d){
 	//	__syncthreads(); to use after reading data into shared memory
-	double density, velocity[D_LBM], feq[Q_LBM], *currentCell;
+	float density, velocity[D_LBM], feq[Q_LBM], *currentCell;
 	int x = 1+threadIdx.x+blockIdx.x*blockDim.x;
 	int y = 1+threadIdx.y+blockIdx.y*blockDim.y;
 	int z = 1+threadIdx.z+blockIdx.z*blockDim.z;
@@ -130,17 +130,17 @@ __global__ void DoColision(double *collide_field_d){
 }
 
 
-void DoCollisionCuda(double *collide_field, int *flag_field, double tau, int xlength){
-	double *collide_field_d=NULL;
+void DoCollisionCuda(float *collide_field, int *flag_field, float tau, int xlength){
+	float *collide_field_d=NULL;
 	int num_cells = pow(xlength+2, D_LBM);
-	size_t collide_field_size = Q_LBM*num_cells*sizeof(double);
+	size_t collide_field_size = Q_LBM*num_cells*sizeof(float);
 
 	//initialize working data
 	cudaErrorCheck(cudaMalloc(&collide_field_d, collide_field_size));
 	cudaErrorCheck(cudaMemcpy(collide_field_d, collide_field, collide_field_size, cudaMemcpyHostToDevice));
 
 	//initialize constant data
-	cudaErrorCheck(cudaMemcpyToSymbol(tau_d, &tau, sizeof(double), 0, cudaMemcpyHostToDevice));
+	cudaErrorCheck(cudaMemcpyToSymbol(tau_d, &tau, sizeof(float), 0, cudaMemcpyHostToDevice));
 	cudaErrorCheck(cudaMemcpyToSymbol(xlength_d, &xlength, sizeof(int), 0, cudaMemcpyHostToDevice));
 
 	//define grid structure
@@ -163,7 +163,7 @@ void DoCollisionCuda(double *collide_field, int *flag_field, double tau, int xle
 /**
  * Performs the actual streaming computation
  */
-__global__ void DoStreaming(double *stream_field_d, double *collide_field_d){
+__global__ void DoStreaming(float *stream_field_d, float *collide_field_d){
 	//	__syncthreads(); to use after reading data into shared memory
 	int x = 1+threadIdx.x+blockIdx.x*blockDim.x;
 	int y = 1+threadIdx.y+blockIdx.y*blockDim.y;
@@ -184,10 +184,10 @@ __global__ void DoStreaming(double *stream_field_d, double *collide_field_d){
 }
 
 
-void DoStreamingCuda(double *collide_field, double *stream_field, int *flag_field, int xlength){
-	double *collide_field_d=NULL, *stream_field_d=NULL;
+void DoStreamingCuda(float *collide_field, float *stream_field, int *flag_field, int xlength){
+	float *collide_field_d=NULL, *stream_field_d=NULL;
 	int num_cells = pow(xlength+2, D_LBM);
-	size_t computational_field_size = Q_LBM*num_cells*sizeof(double);
+	size_t computational_field_size = Q_LBM*num_cells*sizeof(float);
 
 	//initialize working data
 	cudaErrorCheck(cudaMalloc(&collide_field_d, computational_field_size));
