+++ tum/computational-fluid-dynamics/lbm-gpu/src/collision_gpu.h
@@ -4,12 +4,12 @@
 /** Carries out the whole local collision process. Computes density and velocity and
  *  equilibrium distributions. Carries out BGK update.
  */
-void DoCollisionCuda(double *collide_field, int *flag_field, double tau, int xlength);
+void DoCollisionCuda(float *collide_field, int *flag_field, float tau, int xlength);
 
 /**
  * Carries out the streaming step and writes the respective distribution functions from
  * collideField to streamField.
  */
-void DoStreamingCuda(double *collide_field, double *stream_field, int *flag_field, int xlength);
+void DoStreamingCuda(float *collide_field, float *stream_field, int *flag_field, int xlength);
 
 #endif
