+++ tum/computational-fluid-dynamics/lbm-gpu/src/initialization.h
@@ -4,8 +4,8 @@
 /** reads the parameters for the lid driven cavity scenario from a config file */
 int ReadParameters(
     int *xlength,                       /* reads domain size. Parameter name: "xlength" */
-    double *tau,                        /* relaxation parameter tau. Parameter name: "tau" */
-    double *velocity_wall,              /* velocity of the lid. Parameter name: "characteristicvelocity" */
+    float *tau,                        /* relaxation parameter tau. Parameter name: "tau" */
+    float *velocity_wall,              /* velocity of the lid. Parameter name: "characteristicvelocity" */
     int *timesteps,                     /* number of timesteps. Parameter name: "timesteps" */
     int *timesteps_per_plotting,        /* timesteps between subsequent VTK plots. Parameter name: "vtkoutput" */
     int argc,                           /* number of arguments. Should equal 2 (program + name of config file */
@@ -14,5 +14,5 @@ int ReadParameters(
 
 
 /* initialises the particle distribution functions and the flagfield */
-void InitialiseFields(double *collide_field, double *stream_field,int *flag_field, int xlength);
+void InitialiseFields(float *collide_field, float *stream_field,int *flag_field, int xlength);
 #endif
