+++ tum/computational-fluid-dynamics/lbm-gpu/src/lbm_definitions.h
@@ -12,18 +12,18 @@
 #define MOVING_WALL 2
 //computation enhancing values
 #define EPS 0.05
-static const double C_S_POW2_INV = 3.0;
-static const double C_S_POW4_INV = 9.0;
+static const float C_S_POW2_INV = 3.0;
+static const float C_S_POW4_INV = 9.0;
 //reference values, not used in actual computation
 #define SQRT3 1.73205080756887729
-static const double C_S = 1.0/SQRT3;
+static const float C_S = 1.0/SQRT3;
 //predefined simulation parameters
 static const int LATTICE_VELOCITIES[19][3] = {
     {0,-1,-1},{-1,0,-1},{0,0,-1},{1,0,-1},{0,1,-1},{-1,-1,0},{0,-1,0},{1,-1,0},
     {-1,0,0}, {0,0,0},  {1,0,0}, {-1,1,0},{0,1,0}, {1,1,0},  {0,-1,1},{-1,0,1},
     {0,0,1},  {1,0,1},  {0,1,1}
 };
-static const double LATTICE_WEIGHTS[19] = {
+static const float LATTICE_WEIGHTS[19] = {
     1.0/36.0, 1.0/36.0, 2.0/36.0, 1.0/36.0, 1.0/36.0, 1.0/36.0, 2.0/36.0, 1.0/36.0, 
     2.0/36.0, 12.0/36.0,2.0/36.0, 1.0/36.0, 2.0/36.0, 1.0/36.0, 1.0/36.0, 1.0/36.0, 
     2.0/36.0, 1.0/36.0, 1.0/36.0
