+++ tum/computational-fluid-dynamics/lbm-gpu/src/streaming.h
@@ -5,5 +5,5 @@
  * Carries out the streaming step and writes the respective distribution functions from
  * collideField to streamField.
  */
-void DoStreaming(double *collide_field, double *stream_field, int *flag_field, int xlength);
+void DoStreaming(float *collide_field, float *stream_field, int *flag_field, int xlength);
 #endif
