+++ tum/computational-fluid-dynamics/lbm-gpu/src/visualization.h
@@ -5,7 +5,7 @@
  *  to a file determined by 'filename' and timestep 't'. You can re-use parts of the code
  *  from visual.c (VTK output for Navier-Stokes solver) and modify it for 3D datasets.
  */
-void WriteVtkOutput(const double * const collide_field, const int * const flag_field, 
+void WriteVtkOutput(const float * const collide_field, const int * const flag_field,
         const char * filename, unsigned int t, int xlength);
 
 /**
@@ -15,9 +15,9 @@ void WriteVtkOutput(const double * const collide_field, const int * const flag_f
  * @param ncell
  *          number of inner cells, the ones which are there before adding a boundary layer
  */
-void PrintField(double *field, int ncell);
+void PrintField(float *field, int ncell);
 
-void WriteField(const double * const field, const char * filename, unsigned int t, const int xlength,
+void WriteField(const float * const field, const char * filename, unsigned int t, const int xlength,
 		const int rank);
 
 #endif
